requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

langchain==0.3.19
langchain-core==0.3.40
langchain-community==0.3.18
langchain-openai==0.3.7
langchain-huggingface==0.1.2
langchain-chroma==0.2.2

chromadb==0.6.3
sentence-transformers==3.4.1
transformers==4.49.0
torch==2.6.0

pymupdf==1.25.3
pdfplumber==0.11.5
pillow==11.1.0
pytesseract==0.3.13
